agentcore-demo/requirements.txt
bedrock-agentcore>=1.1.0
strands-agents>=0.1.0
strands-agents-tools>=0.1.0